updated to general camera component
