+++ libs/expo/shared/ui-components/src/lib/LibraryModal/index.tsx
@@ -6,9 +6,9 @@ import * as ImagePicker from 'expo-image-picker';
 import { useState } from 'react';
 import { Alert, Pressable, View } from 'react-native';
 import Modal from 'react-native-modal';
+import LibCamera from '../Camera';
 import TextBold from '../TextBold';
 import TextRegular from '../TextRegular';
-import LibCamera from './Camera';
 
 interface ILibraryModalProps {
   onCapture: (file: ReactNativeFile) => void;
