+++ libs/expo/shared/ui-components/src/lib/index.ts
@@ -6,7 +6,7 @@ export { default as BasicRadio } from './BasicRadio';
 export { default as BasicTextarea } from './BasicTextarea';
 export { default as BottomActions } from './BottomActions';
 export { default as Button } from './Button';
-export { default as CameraPicker } from './CameraPicker';
+export { default as Camera } from './Camera';
 export { default as Checkbox } from './Checkbox';
 export { default as CircularLoading } from './CircularLoading';
 export { default as ClientCard } from './ClientCard';
@@ -18,6 +18,7 @@ export { default as FieldCard } from './FieldCard';
 export { default as IconButton } from './IconButton';
 export { default as ImagePicker } from './ImagePicker';
 export { default as Input } from './Input';
+export { default as InteractionCamera } from './InteractionCamera';
 export { default as LibraryModal } from './LibraryModal';
 export { default as Loading } from './Loading';
 export { default as Radio } from './Radio';
